An assortment of patches that reduce our incompatibilities with the upcoming gcc 13.

Closes #13243

* github.com:scylladb/scylladb:
  transport: correctly format unknown opcode
  treewide: catch by reference
  test: raft: avoid confusing string compare
  utils, types, test: extract lexicographical compare utilities
  test: raft: fsm_test: disambiguate raft::configuration construction
  test: reader_concurrency_semaphore_test: handle all enum values
  repair: fix signed/unsigned compare
  repair: fix incorrect signed/unsigned compare
  treewide: avoid unused variables in if statements
  keys: disambiguate construction from initializer_list<bytes>
  cql3: expr: fix serialize_listlike() reference-to-temporary with gcc
  compaction: error on invalid scrub type
  treewide: prevent redefining names
  api: task_manager: fix signed/unsigned compare
  alternator: streams: fix signed/unsigned comparison
  test: fix some mismatched signed/unsigned comparisons

+++ test/boost/cached_file_test.cc
@@ -315,7 +315,7 @@ SEASTAR_THREAD_TEST_CASE(test_stress_eviction) {
         return cf_lru.evict();
     });
 
-    for (int i = 0; i < (cached_size / page_size); ++i) {
+    for (size_t i = 0; i < (cached_size / page_size); ++i) {
         read_to_string(cf, page_size * i, page_size);
     }
 
@@ -335,7 +335,7 @@ SEASTAR_THREAD_TEST_CASE(test_stress_eviction) {
     testlog.debug("Memory: allocated={}, free={}", seastar::memory::stats().allocated_memory(), seastar::memory::stats().free_memory());
     testlog.debug("Starting test...");
 
-    for (int j = 0; j < n_pages * 16; ++j) {
+    for (size_t j = 0; j < n_pages * 16; ++j) {
         testlog.trace("Allocating");
         auto stride = tests::random::get_int(1, 20);
         auto page_idx = tests::random::get_int(n_pages - stride);
